content-generator-function/requirements.txt
functions-framework>=3.0.0
google-cloud-storage>=2.0.0
google-cloud-pubsub>=2.13.0
